extract kernels into separate module to avoid cyclic imports
