+++ torchvision/prototype/__init__.py
@@ -1,1 +1,1 @@
-from . import datasets, features, models, transforms, utils
+from . import constants, datasets, features, kernels, models, transforms, utils
