+++ torchvision/prototype/features/__init__.py
@@ -1,6 +1,7 @@
-from ._bounding_box import BoundingBox, BoundingBoxFormat
+from ._feature import _Feature  # usort: skip
+
+from ._bounding_box import BoundingBox
 from ._encoded import EncodedData, EncodedImage, EncodedVideo
-from ._feature import _Feature
-from ._image import ColorSpace, Image
+from ._image import Image
 from ._label import Label, OneHotLabel
 from ._segmentation_mask import SegmentationMask
