+++ torchvision/prototype/kernels/__init__.py
@@ -1,0 +1,7 @@
+from ._meta import convert_bounding_box_format  # usort:skip
+from ._geometry import (
+    horizontal_flip_bounding_box,
+    horizontal_flip_image_pil,
+    horizontal_flip_image_tensor,
+    horizontal_flip_segmentation_mask,
+)
