+++ torchvision/prototype/transforms/_meta.py
@@ -2,7 +2,8 @@ from typing import Any, Dict, Optional, Union
 
 import PIL.Image
 import torch
-from torchvision.prototype import features
+from torchvision.prototype import features, kernels as K
+from torchvision.prototype.constants import BoundingBoxFormat, ColorSpace
 from torchvision.prototype.transforms import functional as F, Transform
 from torchvision.transforms.functional import convert_image_dtype
 
@@ -10,15 +11,15 @@ from ._utils import is_simple_tensor
 
 
 class ConvertBoundingBoxFormat(Transform):
-    def __init__(self, format: Union[str, features.BoundingBoxFormat]) -> None:
+    def __init__(self, format: Union[str, BoundingBoxFormat]) -> None:
         super().__init__()
         if isinstance(format, str):
-            format = features.BoundingBoxFormat[format]
+            format = BoundingBoxFormat[format]
         self.format = format
 
     def _transform(self, inpt: Any, params: Dict[str, Any]) -> Any:
         if isinstance(inpt, features.BoundingBox):
-            output = F.convert_bounding_box_format(inpt, old_format=inpt.format, new_format=params["format"])
+            output = K.convert_bounding_box_format(inpt, old_format=inpt.format, new_format=params["format"])
             return features.BoundingBox.new_like(inpt, output, format=params["format"])
         else:
             return inpt
@@ -42,17 +43,17 @@ class ConvertImageDtype(Transform):
 class ConvertImageColorSpace(Transform):
     def __init__(
         self,
-        color_space: Union[str, features.ColorSpace],
-        old_color_space: Optional[Union[str, features.ColorSpace]] = None,
+        color_space: Union[str, ColorSpace],
+        old_color_space: Optional[Union[str, ColorSpace]] = None,
     ) -> None:
         super().__init__()
 
         if isinstance(color_space, str):
-            color_space = features.ColorSpace.from_str(color_space)
+            color_space = ColorSpace.from_str(color_space)
         self.color_space = color_space
 
         if isinstance(old_color_space, str):
-            old_color_space = features.ColorSpace.from_str(old_color_space)
+            old_color_space = ColorSpace.from_str(old_color_space)
         self.old_color_space = old_color_space
 
     def _transform(self, inpt: Any, params: Dict[str, Any]) -> Any:
